Player Control Toggle
When using your bot, I find it annoying if the player control keeps posting to a certain channel. So I made a new feature called `Player Control Toggle` that allows us to turn player control on or off at any time. This is made by me and made by myself, so please accept this pull request of mine!